Move shields button to URL Bar (from inside the text box)
