+++ Client/Frontend/Browser/TabLocationView.swift
@@ -14,7 +14,6 @@ protocol TabLocationViewDelegate {
     func tabLocationViewDidTapLocation(_ tabLocationView: TabLocationView)
     func tabLocationViewDidLongPressLocation(_ tabLocationView: TabLocationView)
     func tabLocationViewDidTapReaderMode(_ tabLocationView: TabLocationView)
-    func tabLocationViewDidTapBraveShieldsButton(_ tabLocationView: TabLocationView)
     func tabLocationViewDidTapPageOptions(_ tabLocationView: TabLocationView, from button: UIButton)
     func tabLocationViewDidLongPressPageOptions(_ tabLocationVIew: TabLocationView)
     func tabLocationViewDidBeginDragInteraction(_ tabLocationView: TabLocationView)
@@ -50,18 +49,7 @@ class TabLocationView: UIView {
             updateLockImageView()
             updateTextWithURL()
             pageOptionsButton.isHidden = (url == nil)
-            refreshShieldsStatus()
             setNeedsUpdateConstraints()
-        }
-    }
-    
-    /// Update the shields icon based on whether or not shields are enabled for this site
-    func refreshShieldsStatus() {
-        if let domain = url?.normalizedHost, let state = BraveShieldState.getStateForDomain(domain),
-            let shieldsAllOffOverride = state.isShieldOverrideEnabled(.AllOff), shieldsAllOffOverride {
-            shieldsButton.setImage(UIImage(imageLiteralResourceName: "shields-off-menu-icon"), for: .normal)
-        } else {
-            shieldsButton.setImage(UIImage(imageLiteralResourceName: "shields-menu-icon"), for: .normal)
         }
     }
 
@@ -172,17 +160,6 @@ class TabLocationView: UIView {
         return pageOptionsButton
     }()
     
-    lazy var shieldsButton: ToolbarButton = {
-        let button = ToolbarButton()
-        button.setImage(UIImage(imageLiteralResourceName: "shields-menu-icon"), for: .normal)
-        button.addTarget(self, action: #selector(tappedBraveShieldsButton), for: .touchUpInside)
-        button.isAccessibilityElement = true
-        button.imageView?.contentMode = .left
-        button.accessibilityLabel = Strings.Brave_Panel
-        button.accessibilityIdentifier = "TabLocationView.shieldsButton"
-        return button
-    }()
-    
     lazy var separatorLine: UIView = {
         let line = UIView()
         line.layer.cornerRadius = 2
@@ -204,7 +181,7 @@ class TabLocationView: UIView {
         addGestureRecognizer(longPressRecognizer)
         addGestureRecognizer(tapRecognizer)
 
-        let subviews = [lockImageView, urlTextField, readerModeButton, separatorLine, pageOptionsButton, shieldsButton]
+        let subviews = [lockImageView, urlTextField, readerModeButton, separatorLine, pageOptionsButton]
         contentView = UIStackView(arrangedSubviews: subviews)
         contentView.distribution = .fill
         contentView.alignment = .center
@@ -222,10 +199,6 @@ class TabLocationView: UIView {
         }
 
         pageOptionsButton.snp.makeConstraints { make in
-            make.size.equalTo(TabLocationViewUX.ButtonSize)
-        }
-        
-        shieldsButton.snp.makeConstraints { make in
             make.size.equalTo(TabLocationViewUX.ButtonSize)
         }
         
@@ -252,15 +225,11 @@ class TabLocationView: UIView {
 
     override var accessibilityElements: [Any]? {
         get {
-            return [lockImageView, urlTextField, readerModeButton, pageOptionsButton, shieldsButton].filter { !$0.isHidden }
+            return [lockImageView, urlTextField, readerModeButton, pageOptionsButton].filter { !$0.isHidden }
         }
         set {
             super.accessibilityElements = newValue
         }
-    }
-    
-    @objc func tappedBraveShieldsButton() {
-        delegate?.tabLocationViewDidTapBraveShieldsButton(self)
     }
 
     @objc func tapReaderModeButton() {
